requirements.txt
pyyaml>=6.0,<7
openai>=1.0,<2
anthropic>=0.30,<1
ollama>=0.3,<1
